move files into chess module
Bench: 5239937
