+++ src/chess/board/mod.rs
@@ -14,10 +14,10 @@ use anyhow::{bail, Context};
 use rand::{prelude::SliceRandom, rngs::ThreadRng};
 
 use crate::{
-    board::movegen::{
+    chess::board::movegen::{
         bishop_attacks, king_attacks, knight_attacks, pawn_attacks, rook_attacks, MoveList,
     },
-    chessmove::Move,
+    chess::chessmove::Move,
     cuckoo,
     historytable::ContHistIndex,
     makemove::{hash_castling, hash_ep, hash_piece, hash_side},
@@ -2078,7 +2078,7 @@ mod tests {
     fn game_end_states() {
         use super::Board;
         use super::{DrawType, GameOutcome};
-        use crate::{chessmove::Move, util::Square};
+        use crate::{chess::chessmove::Move, util::Square};
 
         let mut fiftymove_draw =
             Board::from_fen("rnbqkb1r/pppppppp/5n2/8/3N4/8/PPPPPPPP/RNBQKB1R b KQkq - 100 2")
@@ -2123,7 +2123,7 @@ mod tests {
 
     #[test]
     fn fen_round_trip() {
-        use crate::board::Board;
+        use crate::chess::board::Board;
         use std::{
             fs::File,
             io::{BufRead, BufReader},
@@ -2175,7 +2175,7 @@ mod tests {
     #[test]
     fn castling_pseudolegality() {
         use super::Board;
-        use crate::chessmove::{Move, MoveFlags};
+        use crate::chess::chessmove::{Move, MoveFlags};
         use crate::util::Square;
         let board =
             Board::from_fen("1r2k2r/2pb1pp1/2pp4/p1n5/2P4p/PP2P2P/1qB2PP1/R2QKN1R w KQk - 0 20")
@@ -2211,7 +2211,7 @@ mod tests {
     #[test]
     fn key_after_works_for_simple_moves() {
         use super::Board;
-        use crate::chessmove::Move;
+        use crate::chess::chessmove::Move;
         use crate::util::Square;
         let mut board = Board::default();
         let mv = Move::new(Square::E2, Square::E3);
@@ -2223,7 +2223,7 @@ mod tests {
     #[test]
     fn key_after_works_for_captures() {
         use super::Board;
-        use crate::chessmove::Move;
+        use crate::chess::chessmove::Move;
         use crate::util::Square;
         let mut board =
             Board::from_fen("r1bqkb1r/ppp2ppp/2n5/3np1N1/2B5/8/PPPP1PPP/RNBQK2R w KQkq - 0 6")
@@ -2247,7 +2247,7 @@ mod tests {
     #[test]
     fn ep_square_edge_case() {
         use super::Board;
-        use crate::chessmove::Move;
+        use crate::chess::chessmove::Move;
         use crate::makemove::{hash_ep, hash_piece, hash_side};
         use crate::piece::Piece;
         use crate::util::Square;
@@ -2281,7 +2281,7 @@ mod tests {
     #[test]
     fn other_ep_edge_case() {
         use super::Board;
-        use crate::chessmove::Move;
+        use crate::chess::chessmove::Move;
         use crate::util::Square;
         let mut board =
             Board::from_fen("rnbqkbnr/1ppppppp/p7/P7/8/8/1PPPPPPP/RNBQKBNR b KQkq - 0 2").unwrap();
