+++ src/tablebases/probe.rs
@@ -6,8 +6,8 @@
 )]
 
 use crate::{
-    board::{evaluation::TB_WIN_SCORE, movegen::MoveList, Board},
-    chessmove::Move,
+    chess::board::{evaluation::TB_WIN_SCORE, movegen::MoveList, Board},
+    chess::chessmove::Move,
     piece::{Colour, PieceType},
     tablebases::bindings::{
         tb_init, tb_probe_root, tb_probe_wdl, TB_BLESSED_LOSS, TB_CURSED_WIN, TB_DRAW, TB_LARGEST,
